@@ -1,90 +1,147 @@
 {
   "name": "omi",
-  "version": "1.8.1",
-  "description": "Open and modern framework for building user interfaces.",
+  "version": "3.0.0",
+  "description": "Omi === Preact + Scoped CSS + Store System + Native Support in 3kb javascript.",
   "main": "dist/omi.js",
+  "jsnext:main": "dist/omi.esm.js",
+  "module": "dist/omi.esm.js",
+  "dev:main": "dist/omi.dev.js",
+  "minified:main": "dist/omi.min.js",
+  "types": "dist/omi.d.ts",
   "scripts": {
-    "lint": "eslint src",
-    "lint:fix": "eslint src --fix",
-    "lint:example": "eslint example --fix",
-    "store": "webpack -w",
-    "simple": "webpack -w",
-    "nest": "webpack -w",
-    "scoped-style": "webpack -w",
-    "on": "webpack -w",
-    "render-to-string": "webpack -w",
-    "newly-added": "webpack -w",
-    "undo": "webpack -w",
-    "slot": "webpack -w",
-    "plugin": "webpack -w",
-    "todo": "webpack -w",
-    "ref": "webpack -w",
-    "tree": "webpack -w",
-    "perfs": "webpack -w",
-    "life-cycle": "webpack -w",
-    "pagination": "webpack -w",
-    "build": "webpack",
-    "hyperscript": "webpack -w",
-    "select": "webpack -w",
-    "build-min": "webpack",
-    "test": "karma start test/karma.conf.js"
+    "clean": "rimraf dist/ devtools.js devtools.js.map debug.js debug.js.map test/ts/**/*.js",
+    "copy-flow-definition": "copyfiles -f src/omi.js.flow dist",
+    "copy-typescript-definition": "copyfiles -f src/omi.d.ts dist",
+    "build": "npm-run-all --silent clean transpile copy-flow-definition copy-typescript-definition strip optimize minify size",
+    "flow": "flow",
+    "todo": "rollup -c config/rollup.example.js --watch",
+    "store": "rollup -c config/rollup.example.js --watch",
+    "omi-tap": "rollup -c config/rollup.example.js --watch",
+    "simple": "rollup -c config/rollup.example.js --watch",
+    "transpile:main": "rollup -c config/rollup.config.js",
+    "transpile:devtools": "rollup -c config/rollup.config.devtools.js",
+    "transpile:esm": "rollup -c config/rollup.config.esm.js",
+    "transpile:debug": "babel debug/ -o debug.js -s",
+    "transpile": "npm-run-all transpile:main transpile:esm transpile:devtools transpile:debug",
+    "optimize": "uglifyjs dist/omi.dev.js -c conditionals=false,sequences=false,loops=false,join_vars=false,collapse_vars=false --pure-funcs=Object.defineProperty --mangle-props --mangle-regex=\"/^(_|normalizedNodeName|nextBase|prev[CPS]|_parentC)/\" --name-cache config/properties.json -b width=120,quote_style=3 -o dist/omi.js -p relative --in-source-map dist/omi.dev.js.map --source-map dist/omi.js.map",
+    "minify": "uglifyjs dist/omi.js -c collapse_vars,evaluate,screw_ie8,unsafe,loops=false,keep_fargs=false,pure_getters,unused,dead_code -m -o dist/omi.min.js -p relative --in-source-map dist/omi.js.map --source-map dist/omi.min.js.map",
+    "strip:main": "jscodeshift --run-in-band -s -t config/codemod-strip-tdz.js dist/omi.dev.js && jscodeshift --run-in-band -s -t config/codemod-const.js dist/omi.dev.js && jscodeshift --run-in-band -s -t config/codemod-let-name.js dist/omi.dev.js",
+    "strip:esm": "jscodeshift --run-in-band -s -t config/codemod-strip-tdz.js dist/omi.esm.js && jscodeshift --run-in-band -s -t config/codemod-const.js dist/omi.esm.js && jscodeshift --run-in-band -s -t config/codemod-let-name.js dist/omi.esm.js",
+    "strip": "npm-run-all strip:main strip:esm",
+    "size": "node -e \"process.stdout.write('gzip size: ')\" && gzip-size --raw dist/omi.min.js",
+    "test": "npm-run-all lint --parallel test:mocha test:karma test:ts test:flow test:size",
+    "fix": "eslint src --fix",
+    "test:flow": "flow check",
+    "test:ts": "tsc -p test/ts/ && mocha --require babel-register test/ts/**/*-test.js",
+    "test:mocha": "mocha --recursive --require babel-register test/shared test/node",
+    "test:karma": "karma start test/karma.conf.js --single-run",
+    "test:mocha:watch": "npm run test:mocha -- --watch",
+    "test:karma:watch": "npm run test:karma -- no-single-run",
+    "test:size": "bundlesize",
+    "lint": "eslint debug devtools src test",
+    "prepublishOnly": "npm run build",
+    "smart-release": "npm run build && npm test && git commit -am $npm_package_version && git tag $npm_package_version && git push && git push --tags && npm publish",
+    "release": "cross-env npm run smart-release"
   },
+  "eslintConfig": {
+    "extends": "./config/eslint-config.js"
+  },
+  "typings": "./dist/omi.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/AlloyTeam/omi.git"
+  },
+  "files": [
+    "devtools",
+    "debug",
+    "src",
+    "dist",
+    "devtools.js",
+    "devtools.js.map",
+    "debug.js",
+    "debug.js.map",
+    "typings.json"
+  ],
+  "keywords": [
+    "omi",
+    "preact",
+    "react",
+    "virtual dom",
+    "vdom",
+    "components",
+    "virtual",
+    "dom"
+  ],
+  "author": "dntzhang <dntzhang@qq.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/AlloyTeam/omi/issues"
+  },
+  "homepage": "https://github.com/AlloyTeam/omi",
   "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
-    "babel-plugin-transform-es3-property-literals": "^6.22.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-omi": "^0.1.0",
+    "@types/chai": "^4.1.2",
+    "@types/mocha": "^5.0.0",
+    "@types/node": "^9.4.7",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-eslint": "^8.2.2",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "babel-preset-env": "^1.6.1",
+    "bundlesize": "^0.17.0",
     "chai": "^4.1.2",
-    "eslint": "^4.3.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.1.1",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "esprima-fb": "^15001.1001.0-dev-harmony-fb",
-    "html2hscript": "^2.0.1",
-    "hyperscript-helpers": "^3.0.3",
-    "jasmine-core": "^2.5.2",
-    "jstransform": "^11.0.3",
-    "karma": "^1.3.0",
+    "copyfiles": "^2.0.0",
+    "core-js": "^2.4.1",
+    "coveralls": "^3.0.0",
+    "cross-env": "^5.1.4",
+    "diff": "^3.0.0",
+    "eslint": "^4.18.2",
+    "eslint-plugin-react": "^7.7.0",
+    "flow-bin": "^0.67.1",
+    "gzip-size-cli": "^2.0.0",
+    "istanbul-instrumenter-loader": "^3.0.0",
+    "jscodeshift": "^0.5.0",
+    "karma": "^2.0.0",
     "karma-babel-preprocessor": "^7.0.0",
     "karma-chai-sinon": "^0.1.5",
     "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.0.0",
-    "karma-jasmine": "^1.1.0",
     "karma-mocha": "^1.1.1",
     "karma-mocha-reporter": "^2.0.4",
     "karma-sauce-launcher": "^1.1.0",
     "karma-sinon": "^1.0.5",
     "karma-source-map-support": "^1.2.0",
     "karma-sourcemap-loader": "^0.3.6",
-    "karma-webpack": "^2.0.4",
+    "karma-webpack": "^3.0.0",
     "mocha": "^5.0.4",
-    "node-libs-browser": "^0.5.3",
-    "sinon": "^4.5.0",
+    "npm-run-all": "^4.0.0",
+    "rimraf": "^2.5.3",
+    "rollup": "^0.57.1",
+    "rollup-plugin-babel": "^3.0.2",
+    "rollup-plugin-license": "^0.6.0",
+    "rollup-plugin-memory": "^3.0.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "sinon": "^4.4.2",
     "sinon-chai": "^3.0.0",
-    "webpack": "^3.4.1"
+    "typescript": "^2.8.1",
+    "uglify-js": "^2.7.5",
+    "webpack": "^4.3.0"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/AlloyTeam/omix.git"
+  "greenkeeper": {
+    "ignore": [
+      "babel-cli",
+      "babel-core",
+      "babel-eslint",
+      "babel-loader",
+      "jscodeshift",
+      "rollup-plugin-babel"
+    ]
   },
-  "keywords": [
-    "omi",
-    "omix"
-  ],
-  "devEngines": {
-    "node": "6.x || 7.x || 8.x",
-    "npm": "3.x || 4.x || 5.x"
-  },
-  "author": "dntzhang",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/AlloyTeam/omix/issues/new"
-  },
-  "homepage": "http://alloyteam.github.io/omix"
+  "bundlesize": [
+    {
+      "path": "./dist/omi.min.js",
+      "threshold": "4Kb"
+    }
+  ]
 }
