@@ -1,6 +1,6 @@
 {
   "name": "omi-finger",
-  "version": "0.2.4",
+  "version": "1.0.0",
   "description": "AlloyFinger / Omi integration. Support touch and gesture events in your Omi project. ",
   "main": "index.js",
   "scripts": {
@@ -15,20 +15,17 @@
     "alloyfinger": "^0.1.10"
   },
   "devDependencies": {
-    "babel-core": "^6.0.20",
-    "babel-loader": "^6.0.1",
-    "babel-preset-es2015": "^6.0.15",
-    "node-libs-browser": "^0.5.3",
-    "virtual-dom-loader": "^0.1.0",
-    "webpack": "^1.14.0",
-    "omix": "latest",
-    "omi": "latest"
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.1",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-omi": "^0.1.1",
+    "esprima-fb": "^15001.1001.0-dev-harmony-fb",
+    "webpack": "^3.4.1"
   },
   "keywords": [
     "omi",
     "omi-finger",
-    "alloyfinger",
-    "omix"
+    "alloyfinger"
   ],
   "author": "dntzhang",
   "license": "MIT",
