@@ -1,29 +1,30 @@
 {
   "name": "omi-transform",
-  "version": "0.3.1",
+  "version": "1.0.0",
   "description": "Omi / css3transform  integration. Support CSS3 transformation in your Omi project. ",
   "main": "index.js",
   "scripts": {
     "simple": "webpack -w",
-    "sync-for-update": "webpack -w",
-    "nest": "webpack -w"
+    "build": "webpack",
+    "build-min": "webpack"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/AlloyTeam/omi.git"
   },
   "dependencies": {
-    "css3transform": "^1.1.8"
+    "css3transform": "^1.1.9"
   },
   "devDependencies": {
-    "babel-core": "^6.0.20",
-    "babel-loader": "^6.0.1",
-    "babel-preset-es2015": "^6.0.15",
-    "node-libs-browser": "^0.5.3",
-    "webpack": "^1.14.0",
-    "virtual-dom-loader": "^0.1.0",
-    "omix": "latest",
-    "omi": "latest"
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
+    "babel-plugin-transform-es3-property-literals": "^6.22.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-omi": "^0.1.1",
+    "esprima-fb": "^15001.1001.0-dev-harmony-fb",
+    "webpack": "^3.4.1"
   },
   "keywords": [
     "omi",
