@@ -1,8 +1,8 @@
-﻿import Omi from 'omi'
+﻿import Omi from '../../../../dist/omi'
 
 class User extends Omi.Component {
 
-    beforeRender(){
+    install(){
         let params =  this.$route.params
         this.name = params.name
         this.category = params.category
@@ -39,6 +39,6 @@ class User extends Omi.Component {
 }
 
 
-Omi.tag('user',User)
+
 
 export default  User
