@@ -8,10 +8,10 @@ var webpack = require('webpack');
 var ENV = process.env.npm_lifecycle_event;
 
 var config  = {
-    entry: './example/' + ENV + '/main.js',
+    entry: __dirname+'/example/' + ENV + '/main.js',
     output: {
         // path: __dirname,
-        path: './example/' + ENV + '/',
+        path: __dirname+'/example/' + ENV + '/',
         filename: 'bundler.js'
     },
     module: {
@@ -20,14 +20,12 @@ var config  = {
                 loader: 'babel-loader',
                 test: /\.js$/,
                 query: {
-                    presets: ['env', 'omi'],
+                    presets: 'env',
                 }
             }
         ]
     },
     plugins: [
-        // Avoid publishing files when compilation fails
-        new webpack.NoErrorsPlugin()
     ],
     stats: {
         // Nice colored output
