@@ -1,6 +1,6 @@
 var path = require('path');
 var webpack = require('webpack');
-
+var packageJSON = require('./package.json');
 /**
  * Env
  * Get npm lifecycle event to identify the environment
@@ -11,7 +11,7 @@ var config  = {
     entry: './example/' + ENV + '/main.js',
     output: {
         // path: __dirname,
-        path: './example/' + ENV + '/',
+        path: path.join(__dirname, './example/' + ENV + '/'),
         filename: 'bundler.js'
     },
     module: {
@@ -20,16 +20,14 @@ var config  = {
                 loader: 'babel-loader',
                 test: /\.js$/,
                 query: {
-                    presets: 'es2015'
+                    presets: 'env'
                 }
-            },{
-                loader: 'virtual-dom-loader?jsx=Omi.x'
             }
         ]
     },
     plugins: [
         // Avoid publishing files when compilation fails
-        new webpack.NoErrorsPlugin()
+        new webpack.NoEmitOnErrorsPlugin()
     ],
     stats: {
         // Nice colored output
@@ -38,5 +36,57 @@ var config  = {
     // Create Sourcemaps for the bundle
     // devtool: 'source-map',
 };
+
+if(ENV === 'build'||ENV === 'build-min'){
+    config = {
+        entry: {
+            'omi-transform': './index.js'
+        },
+        output: {
+            // path: __dirname,
+            path: path.join(__dirname, 'dist/'),
+            library:'OmiTransform',
+            libraryTarget: 'umd',
+            filename:  '[name].js'
+            //umdNamedDefine: true
+        },
+        externals: {
+            "omi": 'omi'
+        },
+        module: {
+            loaders: [
+                {
+                    loader: 'babel-loader',
+                    test: path.join(__dirname, 'src'),
+                    query: {
+                        presets: 'env',
+                        plugins : [
+                            "transform-es3-property-literals",
+                            "transform-es3-member-expression-literals"
+                        ]
+                    },
+                }
+            ]
+        },
+        plugins: [
+            // Avoid publishing files when compilation fails
+            new webpack.BannerPlugin(" OmiRouter v"+packageJSON.version+" By dntzhang \r\n Github: https://github.com/AlloyTeam/omi\r\n MIT Licensed."),
+            new webpack.NoEmitOnErrorsPlugin()
+        ],
+        stats: {
+            // Nice colored output
+            colors: true
+        },
+        // Create Sourcemaps for the bundle
+        // devtool: 'source-map',
+    };
+
+    if(ENV === 'build-min'){
+        config.plugins.push(new webpack.optimize.UglifyJsPlugin());
+        config.entry = {
+            'omi-transform.min': './index.js'
+        };
+    }
+}
 
 module.exports = config;
